Add error stories to react
